[dev] refactor into features/expand_cols
